Bluetooth: Host: Bitwise-or CCC values in `_bt_gatt_ccc.value`
Summarize the value of the CCC using bitwise-or instead of max. This
is more natural and affords distinguishing between when only indications
are enabled and when both notifications and indications are enabled.

It is not expected that this change will affect any applications. It will
not affect attributes that allow only one of notifications and indications.
It will also not affect applications that only compare this value to zero.

But it simplifies the interface of `_bt_gatt_ccc.value` and
`_bt_gatt_ccc.cfg_changed`.

Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>