Bluetooth: Host: Bitwise-or CCC values in `_bt_gatt_ccc.value`
After this change, we summarize the value of the CCC using bitwise-or
instead of max. This is more natural and affords distinguishing between
when only indications are enabled and when both notifications and
indications are enabled.

It is not expected that this change will affect any applications. It will
not affect attributes that allow only one of notifications and indications.
It will also not affect applications that only compare this value to zero.

This does however simplify the API of `_bt_gatt_ccc.value` and
`_bt_gatt_ccc.cfg_changed` by the virtue of being more natural. And
since we are altering the behavior of the API already in the previous
commit, we might as well throw this in.

Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>