Bluetooth: Host: sched_lock for thread safety
Before this change, the Bluetooth host was potentially relying on the
system work queue running in cooperative mode. This was enforced, so
there is bug here.

But the system work queue priority should be configurable by the user
using `CONFIG_SYSTEM_WORKQUEUE_PRIORITY` to fit the need of the
application.  Unfortunately, zephyr enforces that cooperative threads
have more urgent priority than preemptible threads.

To remedy this situation and make priorities more flexible, we remove
the cooperativeness requirements from the host by instead using a
scheduler lock, which is a temporary cooperative setting for a thread.

The locks in this patch are more coarse grained than they may be. Not
much time has been spent analysing atomicity requirements. The changes
are pretty much mechanically done: Any work handler in the host that is
not optionally scheduled on the preemptible long work queue, and thus
presumably thread-safe, gets a surrounded with a scheduler lock.

Feel free refine the sheduler lock placements or remove them if they are
unnecessary.

The TX thread is also given this treatment.

Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>