Bluetooth: service: nus: Switch to one-time adv
The nus service provides a one-click solution for automatically starting
the Bluetooth subsystem and an advertiser.

This patch removes the reliance on the auto-resume feature, which is
about to be deprecated. This requires a new thread to drive the blocking
function `bt_le_adv_start`, since blocking functions should not be
invoked in Bluetooth callbacks.

Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>