task_wdt: Add auto init and syswq monitoring
This patch adds two new features to the task watchdog subsystem:
  - Automatic initialization of the task watchdog on SYS_INIT
  - Monitoring of the system work queue

The automatic initialization feature is enabled by the new Kconfig
option TASK_WDT_AUTO_INIT. When enabled, the task watchdog will be
initialized automatically on SYS_INIT. This is useful for applications
that want to use the task watchdog but do not want to manually
initialize it.

The system work queue monitoring feature is enabled by the new Kconfig
option TASK_WDT_SYSWQ_STALL. When enabled, the task watchdog will
monitor the system work queue. If the system work queue is stalled for
longer than the specified timeout, the task watchdog will trigger a
system panic.

The system work queue stall timeout is specified by the new Kconfig
option TASK_WDT_SYSWQ_STALL_TIMEOUT_MS. This option specifies the
maximum time in milliseconds that the system work queue is allowed to be
stalled before the task watchdog triggers a system panic.

Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>